Updated jobs to be tasks in anticipation of scheduler jobs classes
